Refactor Utility
