+++ Descension/Assets/Scripts/Actor/AI/AIController.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
+using Util.Helpers;
 using UnityEngine;
 using Util.Enums;
+using Util.Constants;
 
 namespace Actor.AI
 {
@@ -113,7 +115,7 @@ namespace Actor.AI
             }
             else
             {
-                _patrolIndex = Util.Util.FindClosest(transform.position, ref patrolTargets);
+                _patrolIndex = CalculationHelper.FindClosest(transform.position, ref patrolTargets);
             }
             Debug.Log(_patrolIndex);
             SetTarget(patrolTargets[_patrolIndex]);
@@ -150,7 +152,7 @@ namespace Actor.AI
         // Look in all directions to see if player is still visible
         private void Look()
         {
-            foreach (var d in Util.Util.Directions)
+            foreach (var d in Direction.Directions)
             {
                 RaycastHit2D rayCast = Physics2D.BoxCast(transform.position, new Vector2(2, 2), 0, d, _attributes.sightDistance, (int) ~UnityLayer.Enemy);
                 if (rayCast)
