+++ Descension/Assets/Scripts/Util/Constants/Direction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 75e8007a3a7788f43b959df6a413ad46
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
