+++ Descension/Assets/Scripts/Util/Helpers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 176979fb8c5aaf543ab5219a545caf0b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
