+++ Descension/Assets/Scripts/Util/Helpers/CalculationHelper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 75853b7bb543b6e4396f2ed35f325728
+guid: 019b5a03a402bf3479ee2c420442eca4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
